version 0.0.63
Long to Int.

+++ src/main/java/guru/springframework/sdjpaintro/dao/BookDaoImpl.java
@@ -26,7 +26,7 @@ public class BookDaoImpl implements BookDao {
             String hql = "SELECT b FROM Book b ORDER BY b.title "
                     + pageable.getSort().getOrderFor("title").getDirection().name();
             TypedQuery<Book> query = em.createQuery(hql, Book.class);
-            query.setFirstResult((int) pageable.getOffset());
+            query.setFirstResult(Math.toIntExact(pageable.getOffset()));
             query.setMaxResults(pageable.getPageSize());
             return query.getResultList();
         }
@@ -36,7 +36,7 @@ public class BookDaoImpl implements BookDao {
     public List<Book> findAllBooks(Pageable pageable) {
         try (EntityManager em = getEntityManager()) {
             TypedQuery<Book> query = em.createQuery("SELECT b FROM Book b", Book.class);
-            query.setFirstResult((int) pageable.getOffset());
+            query.setFirstResult(Math.toIntExact(pageable.getOffset()));
             query.setMaxResults(pageable.getPageSize());
             return query.getResultList();
         }
